recent debugged code
