+++ RDSExpoApp/components/sagas/requests/fetchUser.ts
@@ -1,5 +1,5 @@
 import axios from 'axios';
-import { urls } from '../../constants/appConstant/url';
+import { urls } from '../../../constants/appConstant/url';
 import { User } from '../../context/type';
 
 export const fetchUserData = async (userId: string): Promise<User | null> => {
