+++ RDSExpoApp/package.json
@@ -16,46 +16,46 @@
   },
   "dependencies": {
     "@expo/vector-icons": "^14.0.2",
-    "@react-native-async-storage/async-storage": "^2.1.0",
+    "@react-native-async-storage/async-storage": "1.23.1",
     "@react-native-community/netinfo": "^11.4.1",
     "@react-native-community/slider": "^4.5.5",
     "@react-navigation/bottom-tabs": "^7.2.0",
     "@react-navigation/native": "^7.0.14",
+    "@react-navigation/stack": "^7.1.1",
     "axios": "^1.7.9",
-    "expo": "~52.0.27",
-    "expo-blur": "~14.0.2",
+    "expo": "~52.0.28",
+    "expo-blur": "~14.0.3",
     "expo-camera": "~16.0.14",
-    "expo-constants": "~17.0.4",
+    "expo-constants": "~17.0.5",
     "expo-device": "~7.0.2",
     "expo-font": "~13.0.3",
     "expo-haptics": "~14.0.1",
-    "expo-linking": "~7.0.4",
+    "expo-linking": "~7.0.5",
     "expo-router": "^4.0.17",
+    "expo-secure-store": "~14.0.1",
     "expo-splash-screen": "~0.29.21",
     "expo-status-bar": "~2.0.1",
-    "expo-symbols": "~0.2.1",
+    "expo-symbols": "~0.2.2",
     "expo-system-ui": "~4.0.7",
     "expo-web-browser": "~14.0.2",
     "moment": "^2.30.1",
     "react": "18.3.1",
     "react-dom": "18.3.1",
-    "react-native": "0.76.6",
+    "react-native": "^0.76.6",
     "react-native-circular-progress-indicator": "^4.4.2",
     "react-native-collapsible-tab-view": "^8.0.0",
     "react-native-date-picker": "^5.0.9",
-    "react-native-device-info": "^14.0.2",
     "react-native-element-dropdown": "^2.12.4",
-    "react-native-encrypted-storage": "^4.0.3",
     "react-native-gesture-handler": "~2.20.2",
     "react-native-image-picker": "^7.2.3",
     "react-native-modal": "^13.0.1",
-    "react-native-pager-view": "^6.7.0",
+    "react-native-pager-view": "6.5.1",
     "react-native-paper": "^5.13.1",
     "react-native-reanimated": "~3.16.1",
     "react-native-redash": "^18.1.3",
     "react-native-safe-area-context": "4.12.0",
     "react-native-screens": "~4.4.0",
-    "react-native-svg": "^15.11.1",
+    "react-native-svg": "15.8.0",
     "react-native-toast-message": "^2.2.1",
     "react-native-walkthrough-tooltip": "^1.6.0",
     "react-native-web": "~0.19.13",
